requirements.txt
streamlit>=1.32.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9

# LangChain core + community + Google Generative AI
langchain>=0.1.16
langchain-community>=0.0.34
langchain-google-genai>=1.0.3

# Vector store
chromadb>=0.4.24

# Document parsers
pypdf>=4.1.0
openpyxl>=3.1.2
pandas>=2.2.1

# Utilities
python-dotenv>=1.0.1
pydantic>=2.6.4
